+++ controls-table/src/jvmMain/kotlin/timeseries/RollingSeries.kt
@@ -55,7 +55,5 @@ public class RollingSeries<T>(
         }
     }
 
-    override fun toString(): String {
-        TODO("Not yet implemented")
-    }
+    override fun toString(): String = "RollingSeries(size=$size, position=$position)"
 }